+++ src/main/java/com/sprint/mission/discodeit/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.response;
 
 import java.time.Instant;
 import java.util.Map;
