+++ src/main/java/com/sprint/mission/discodeit/dto/response/PageResponse.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.dto;
+package com.sprint.mission.discodeit.dto.response;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
