+++ src/main/java/com/sprint/mission/discodeit/service/MessageService.java
@@ -2,9 +2,8 @@ package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.dto.MessageDto;
 import com.sprint.mission.discodeit.dto.MessageDto.CreateCommand;
-import com.sprint.mission.discodeit.dto.MessageDto.Detail;
 import com.sprint.mission.discodeit.dto.MessageDto.UpdateCommand;
-import com.sprint.mission.discodeit.dto.PageResponse;
+import com.sprint.mission.discodeit.dto.response.PageResponse;
 import java.time.Instant;
 import java.util.UUID;
 import org.springframework.data.domain.Pageable;
