+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -4,7 +4,7 @@ import com.sprint.mission.discodeit.dto.BinaryContentDto;
 import com.sprint.mission.discodeit.dto.MessageDto;
 import com.sprint.mission.discodeit.dto.MessageDto.CreateCommand;
 import com.sprint.mission.discodeit.dto.MessageDto.UpdateCommand;
-import com.sprint.mission.discodeit.dto.PageResponse;
+import com.sprint.mission.discodeit.dto.response.PageResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Channel;
 import com.sprint.mission.discodeit.entity.Message;
